requirements.txt
python-chess>=1.999
torch>=2.0
numpy
tqdm
h5py
